asan: correct more problems...

+++ src/game/boe.dlgutil.cpp
@@ -549,7 +549,8 @@ void set_up_shop_array() {
 				break;
 		}
 	}
-	shop_sbar->setMaximum(shop_array.size() - 8);
+	// ASAN undefined behaviour
+	shop_sbar->setMaximum(long(shop_array.size()) - 8);
 }
 
 void start_talk_mode(short m_num,short personality,mon_num_t monst_type,short store_face_pic) {
